Created listing page and index to display prediction.
